feat: end game, refactor: useCanvas
