+++ packages/client/src/assets/styles/index.scss
@@ -23,6 +23,7 @@
   --light: #ffffff;
   --dark: #1F252D;
   --silver: #a5aab2;
+  --field-bg: #23272F;
 }
