--- packages/client/src/pages/Game/EndGameModal/EndModal.module.scss
@@ -1,45 +1,0 @@
-.overlay {
-  position: absolute;
-  top: 0; //todo: отрефакторить, когда сделаем компонент модалки с порталом
-  right: 0;
-  left:  0;
-  bottom:  0;
-  background-color: rgba(0, 0, 0, 0.8);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.modal {
-  width: 50%;
-  max-width: 500px;
-  background-color: white;
-  padding: 30px;
-  border-radius: 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-}
-
-.text {
-  font-size: 24px;
-  font-weight: 700;
-}
-
-.buttons {
-  display: flex;
-  gap: 12px;
-}
-
-.button {
-  margin: 0;
-  padding: 10px 20px;
-  font-size: 20px;
-  display: flex;
-  width: fit-content;
-  min-width: 184px;
-  align-items: center;
-  justify-content: center;
-}
