+++ packages/client/src/pages/Game/Game.module.scss
@@ -6,7 +6,7 @@
   grid-template-columns: 3fr 1fr;
   /* width: calc(100% - 128px); */
   /* height: calc(100% - 96px); */
-  /* gap: 24px; */
+  gap: 24px;
 }
 
 .field {
@@ -14,20 +14,23 @@
   display: flex;
   align-items: center;
   justify-content: center;
+  position: relative;
 }
 
-.cards {
-  display: grid;
-  grid-template-columns: repeat(4, 130px);
-  grid-template-rows: repeat(4, 130px);
-  gap: 24px;
-  list-style: none;
-}
-
-.card {
+.endGame {
+  position: absolute;
+  top: 85px;
+  left: 0;
+  right: 0;
+  bottom: 85px;
+  background-color: var(--field-bg);
+  color: var(--light);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 64px;
+  font-weight: 700;
   border-radius: 12px;
-  padding: 20px;
-  background: var(--gray-card);
 }
 
 .handlers {
@@ -81,6 +84,11 @@
   flex-direction: column;
   gap: 12px;
   width: 100%;
+}
+
+.restartButton {
+  font-size: 20px;
+  margin: 0;
 }
 
 .button {
